% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{add.meta.pred}
\alias{add.meta.pred}
\title{Add metadata as predictors}
\usage{
add.meta.pred(feat, meta, pred.names, std.meta)
}
\arguments{
\item{feat}{features object}

\item{meta}{metadata object}

\item{pred.names}{vector of names of the metavariables to be added to the feature matrix as predictors}

\item{std.meta}{Shall added (metadata) features be standardized?}
}
\value{
features object
}
\description{
Adds medata as columns to the feat data to be later used as predictors
}
