% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_utils.r
\name{appendDirName}
\alias{appendDirName}
\title{Append source directory path}
\usage{
appendDirName(source.dir)
}
\arguments{
\item{source.dir}{string with path to the source directory}
}
\value{
string with / at the end
}
\description{
Append / at the end of the name of the directory path if it is not there
}
