% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{associations<-}
\alias{associations<-}
\alias{assign-associations}
\alias{associations<-,siamcat,list-method}
\title{Assign a new assoc object to \code{x}}
\usage{
associations(x) <- value

\S4method{associations}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class list}
}
\value{
none
}
\description{
Assign a new assoc object to \code{x}
}
