% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{associations<-}
\alias{associations<-}
\alias{assign-associations}
\alias{associations<-,siamcat,associations-method}
\title{Assign a new assocications object to \code{x}}
\usage{
associations(x) <- value

\S4method{associations}{siamcat,associations}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an associations object}
}
\value{
none
}
\description{
Assign a new assocications object to \code{x}
}
\examples{
data(siamcat_example)
assocications(siamcat_example) <- associations(siamcat_example)
}
