% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{data_split<-}
\alias{data_split<-}
\alias{assign-data_split}
\alias{data_split<-,siamcat,list-method}
\title{Assign a new list containing a cross-validation split to a
SIAMCAT object}
\usage{
data_split(x) <- value

\S4method{data_split}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{list containing a cross-validation split}
}
\value{
none
}
\description{
Assign a new list containing a cross-validation split to a
SIAMCAT object
}
\examples{
data(siamcat_example)
data_split(siamcat_example) <- data_split(siamcat_example)
}
