% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{features<-}
\alias{features<-}
\alias{assign-features}
\alias{features<-,siamcat,otu_table-method}
\alias{features<-}
\title{Assign a new otu_table object to \code{x} features slot}
\usage{
features(x) <- value

\S4method{features}{siamcat,otu_table}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link[phyloseq]{otu_table-class}}
}
\value{
none
}
\description{
Assign a new otu_table object to \code{x} features slot
}
\examples{
data(siamcat_example)
features(siamcat_example) <- features(siamcat_example)
}
