% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{filt_feat<-}
\alias{filt_feat<-}
\alias{assign-filt_feat}
\alias{filt_feat<-,siamcat,list-method}
\title{Assign a new filt_feat object to \code{x}}
\usage{
filt_feat(x) <- value

\S4method{filt_feat}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an filt_feat object}
}
\value{
none
}
\description{
Assign a new filt_feat object to \code{x}
}
\examples{
data(siamcat_example)
filt_feat(siamcat_example) <- list(
    filt.feat=filt_feat(siamcat_example),
    filt.param=filt_params(siamcat_example))
}
