% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{label<-}
\alias{label<-}
\alias{assign-label}
\alias{label<-,siamcat,list-method}
\title{Assign a new label object to a SIAMCAT object}
\usage{
label(x) <- value

\S4method{label}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an list (in label format)}
}
\value{
none
}
\description{
Assign a new label object to a SIAMCAT object
}
\examples{
data(siamcat_example)
label(siamcat_example) <- label(siamcat_example)
}
