% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{meta<-}
\alias{meta<-}
\alias{assign-meta}
\alias{meta<-,siamcat,sample_data-method}
\alias{meta<-}
\title{Assign a new sam_data object to \code{x}}
\usage{
meta(x) <- value

\S4method{meta}{siamcat,sample_data}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link[phyloseq]{sample_data-class}}
}
\value{
none
}
\description{
Assign a new sam_data object to \code{x}
}
\examples{
data(siamcat_example)
meta(siamcat_example) <- meta(siamcat_example)
}
