% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{model_list<-}
\alias{model_list<-}
\alias{assign-model_list}
\title{Assign a new list containing trained models to a SIAMCAT object}
\usage{
model_list(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{list containing trained models, type of models and of features}
}
\value{
none
}
\description{
Assign a new list containing trained models to a SIAMCAT object
}
\examples{
data(siamcat_example)
model_list(siamcat_example) <- model_list(siamcat_example)
}
