% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{norm_feat<-}
\alias{norm_feat<-}
\alias{assign-norm_feat}
\alias{norm_feat<-,siamcat,norm_feat-method}
\title{Assign a new norm_feat object to \code{x}}
\usage{
norm_feat(x) <- value

\S4method{norm_feat}{siamcat,norm_feat}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an norm_feat object}
}
\value{
none
}
\description{
Assign a new norm_feat object to \code{x}
}
\examples{
data(siamcat_example)
norm_feat(siamcat_example) <- new("norm_feat",
     norm.feat=norm_feat(siamcat_example),
     norm.param=norm_params(siamcat_example))
}
