% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{norm_param<-}
\alias{norm_param<-}
\alias{assign-norm_param}
\alias{norm_param<-,siamcat,list-method}
\alias{norm_param<-}
\title{Assign a new norm_param object to \code{x}}
\usage{
norm_param(x) <- value

\S4method{norm_param}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an norm_param list}
}
\value{
none
}
\description{
Assign a new norm_param object to \code{x}
}
\examples{
data(siamcat_example)
norm_param(siamcat_example) <- norm_param(siamcat_example)
}
