% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{ordination<-}
\alias{ordination<-}
\alias{assign-ordination}
\alias{ordination<-,siamcat,list-method}
\title{Assign a new ordination to a SIAMCAT object}
\usage{
ordination(x) <- value

\S4method{ordination}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{a list of ordination data (see make.ordination)}
}
\value{
none
}
\description{
Assign a new ordination to a SIAMCAT object
}
\examples{
data(siamcat_example)
ordination(siamcat_example) <- ordination(siamcat_example)
}
\keyword{internal}
