% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{phyloseq<-}
\alias{phyloseq<-}
\alias{assign-phyloseq}
\alias{phyloseq<-,siamcat,phyloseq-method}
\alias{phyloseq<-}
\title{Assign a new phyloseq object to \code{x}}
\usage{
phyloseq(x) <- value

\S4method{phyloseq}{siamcat,phyloseq}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link[phyloseq]{phyloseq-class}}
}
\description{
Assign a new phyloseq object to \code{x}
}
\examples{
# data(siamcat_example)
}
