% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{physeq<-}
\alias{physeq<-}
\alias{assign-physeq}
\alias{physeq<-,siamcat,phyloseq-method}
\alias{physeq<-}
\alias{physeq<-,siamcat,otu_table-method}
\alias{physeq<-}
\title{Assign a new phyloseq object to \code{x}}
\usage{
physeq(x) <- value

\S4method{physeq}{siamcat,phyloseq}(x) <- value

\S4method{physeq}{siamcat,otu_table}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link[phyloseq]{phyloseq-class}}
}
\value{
none
}
\description{
Assign a new phyloseq object to \code{x}
}
\examples{
data(siamcat_example)
physeq(siamcat_example) <- physeq(siamcat_example)
}
