% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{pred_matrix<-}
\alias{pred_matrix<-}
\alias{assign-pred_matrix}
\alias{pred_matrix<-,siamcat,matrix-method}
\alias{pred_matrix<-}
\title{Assign a new pred_matrix object to \code{x}}
\usage{
pred_matrix(x) <- value

\S4method{pred_matrix}{siamcat,matrix}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an pred_matrix matrix}
}
\value{
none
}
\description{
Assign a new pred_matrix object to \code{x}
}
\examples{
data(siamcat_example)
pred_matrix(siamcat_example) <- pred_matrix(siamcat_example)
}
