% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{associations}
\alias{associations}
\alias{associations,ANY-method}
\alias{associations,list-method}
\alias{assoc}
\alias{assoc_param}
\alias{assoc_param,ANY-method}
\title{Retrieve associations from object.}
\usage{
associations(siamcat)

\S4method{associations}{ANY}(siamcat, verbose = 1)

\S4method{associations}{list}(siamcat)

assoc_param(siamcat)

\S4method{assoc_param}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains an object in the assoc slot}

\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains an object in the assoc slot}
}
\value{
The results of the association testing or NULL.

The parameters of the assocation testing or NULL
}
\description{
Retrieve associations from object.

Retrieve parameters of association testing from object.
}
\examples{
    data(siamcat_example)
    associations(siamcat_example)
    data(siamcat_example)
    assoc_param(siamcat_example)
}
