% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{assoc_param}
\alias{assoc_param}
\alias{assoc_param,ANY-method}
\alias{assoc_param_param,ANY-method}
\title{Retrieve parameters of association testing from object.}
\usage{
assoc_param(siamcat, verbose=1)

\S4method{assoc_param}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains an object in the associations slot}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The parameters of the assocation testing or NULL
}
\description{
Retrieve parameters of association testing from object.
}
\examples{
    data(siamcat_example)
    assoc_param(siamcat_example)
}
