% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{associations-class}
\alias{associations-class}
\title{The S4 class for storing the results of the association testing}
\description{
The S4 class for storing the results of the association testing
}
\section{Slots}{

\describe{
\item{\code{assoc.results}}{a data.frame containing the results of the
association testing}

\item{\code{assoc.param}}{a list containing the parameters for the association
testing}
}}

