% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_definitions.R
\docType{class}
\name{data_split-class}
\alias{data_split-class}
\title{The S4 class for storing data splits}
\description{
The S4 class for storing data splits
}
\section{Slots}{

\describe{
\item{\code{training.folds}}{a list - for each cv fold contains ids of
samples used for training}

\item{\code{test.folds}}{a list - for each cv fold contains ids of
samples used for testing}

\item{\code{num.resample}}{number of repetition rounds for cv}

\item{\code{num.folds}}{number of folds for cv}
}}

