% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{data_split}
\alias{data_split}
\alias{data_split,ANY-method}
\alias{data_split,data_split-method}
\title{Retrieve a \link{data_split-class} object from object.}
\usage{
label(siamcat, errorIfNULL=TRUE)

\S4method{data_split}{ANY}(siamcat)

\S4method{data_split}{data_split}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link{data_split-class}.}
}
\value{
The \link{data_split-class} object or NULL.
}
\description{
Retrieve a \link{data_split-class} object from object.
}
\examples{
 data(siamcat_example)
 data_split(siamcat_example)
}
