% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_predictions.r
\name{eval.predictions}
\alias{eval.predictions}
\title{Evaluate prediction results}
\usage{
eval.predictions(siamcat)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}
}
\value{
list containing \itemize{
 \item \code{$roc.average} average ROC-curve across repeats or a single ROC-curve on complete dataset;
 \item \code{$auc.average} AUC value for the average ROC-curve;
 \item \code{$ev.list} list of \code{length(num.folds)}, containing for different decision thresholds the number of false positives, false negatives, true negatives, and true positives;
 \item \code{$pr.list} list of \code{length(num.folds)}, containing the positive predictive value (precision) and true positive rate (recall) values used to plot the PR curves;
}. If \code{prediction} had more than one column, i.e. if the models has been trained with several repeats, the function will additonally return \itemize{
 \item \code{$roc.all} list of roc objects (see \link[pROC]{roc}) for every repeat;
 \item \code{$aucspr} vector of AUC values for the PR curves for every repeat;
 \item \code{$auc.all} vector of AUC values for the ROC curves for every repeat
}
}
\description{
This function takes the correct labels and predictions for all samples and evaluates the results using the \itemize{
 \item Area Under the Receiver Operating Characteristic (ROC) Curve (AU-ROC)
 \item and the Precision-recall Curve (PR)
}
as metric. Predictions can be supplied either for a single case or as matrix after resampling of the dataset.

Prediction results are usually produced with the function \link{plm.predictor}.
}
\keyword{SIAMCAT}
\keyword{eval.result}
