% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{eval.result}
\alias{eval.result}
\title{This function splits a dataset into a training and a test set.}
\usage{
eval.result(label, pred)
}
\arguments{
\item{label}{label input file}
}
\description{
This function splits a dataset into a training and a test set.
}
\examples{
eval.result()
}
\keyword{SIAMCAT}
\keyword{association_check}
