% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{eval_data-class}
\alias{eval_data-class}
\title{The S4 class for storing evaluation data.}
\description{
The S4 class for storing evaluation data.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{inherited from \code{\link{list}} class, contains
a list with:
    \itemize{
    \item \code{$roc} average ROC-curve across repeats or a single
    ROC-curve on the complete dataset (object of class \link[pROC]{roc});
    \item \code{$auroc} AUC value for the average ROC-curve;
    \item \code{$prc} average Precision Recall curve across repeats or
    a single PR-curve on the complete dataset;
    \item \code{$auprc} AUC value for the average PR-curve;
    \item \code{$ev} list containing for different decision thresholds the
    number of false positives, false negatives, true negatives, and
    true positives;
}. If \code{prediction} had more than one column, i.e. if the models has
    been trained with several repeats, the function will additonally
    return \itemize{
    \item \code{$roc.all} list of roc objects (see \link[pROC]{roc}) for
    every repeat;
    \item \code{$auroc.all} vector of AUC values for the ROC curves for
    every repeat;
    \item \code{$prc.all} list of PR curves for every repeat;
    \item \code{$auprc.all} vector of AUC values for the PR curves for every
    repeat;
    \item \code{$ev.all} list of false positive, false negatives, true
    negatives, true positives, and thresholds for the different repeats.
}}
}}

