% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_evaluation.r
\name{evaluation.model.plot}
\alias{evaluation.model.plot}
\title{Model Evaluation Plot}
\usage{
evaluation.model.plot(siamcat, fn.plot)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{fn.plot}{string, filename for the pdf-plot}
}
\value{
Does not return anything, but produces the model evaluation plot.
}
\description{
Produces two plots for model evaluation. The first plot shows the Receiver Operating Receiver (ROC)-curves, the other the Precision-recall (PR)-curves for the different CV repetitions.
}
\keyword{SIAMCAT}
\keyword{evaluation.model.plot}
