% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{feat.crc.zeller}
\alias{feat.crc.zeller}
\title{Documentation for the example feature object in the data folder}
\description{
Feature matrix (as data.frame) for the CRC dataset, containing 141 samples
and 1754 bacterial species (features).
}
\keyword{data}
