% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,ANY-method}
\title{Retrieve the active \link[phyloseq]{otu_table-class} out of SIAMCAT}
\usage{
features(siamcat)

\S4method{features}{ANY}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link[phyloseq]{otu_table-class} .}
}
\value{
The \link[phyloseq]{otu_table-class} object or NULL.
}
\description{
Retrieve the active \link[phyloseq]{otu_table-class} out of SIAMCAT
}
\examples{
    data(siamcat_example)
    features(siamcat_example)
}
