% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{filt_feat-class}
\alias{filt_feat-class}
\title{The S4 class for storing the filter features/paramters}
\description{
The S4 class for storing the filter features/paramters
}
\section{Slots}{

\describe{
\item{\code{filt.feat}}{An object of class \link[phyloseq]{otu_table-class}
storing the filtered features}

\item{\code{filt.param}}{A list storing the parameters of the feature filtering}
}}

