% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{filt_feat}
\alias{filt_feat}
\alias{filt_feat,ANY-method}
\title{Retrieve filtered features form object}
\usage{
filt_feat(siamcat, verbose=1)

\S4method{filt_feat}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains an object in the filt_feat slot}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The filtered feature matrix or NULL.
}
\description{
Retrieve filtered features form object
}
\examples{
    data(siamcat_example)
    filt_feat(siamcat_example)
}
