% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_feat.r
\name{filter.feat}
\alias{filter.feat}
\title{Perform unsupervised feature filtering.}
\usage{
filter.feat(siamcat, filter.method = abundance, cutoff = 0.001,
  recomp.prop = FALSE, rm.unmapped = TRUE)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat}}

\item{filter.method}{method used for filtering the features, can be one of these: \code{c("abundance", "cum.abundance", "prevalence")}}

\item{cutoff}{float, abundace or prevalence cutoff}

\item{recomp.prop}{boolean, should absolute abundances be converted into relative abundances?}

\item{rm.unmapped}{boolean, should unmapped reads be discarded?}
}
\value{
siamcat an object of class \link{siamcat}
}
\description{
This function may convert absolute abundances into relative abundances and then performs unsupervised feature filtering. Features can be filtered based on abundance or prevalence. Additionally, unmapped reads may be removed.
}
\keyword{SIAMCAT}
\keyword{filter.feat}
