% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.component.classes}
\alias{get.component.classes}
\title{Show the component objects classes and slot names.}
\usage{
get.component.classes(class)
}
\value{
list of component classes
}
\description{
Show the component objects classes and slot names.
}
\keyword{internal}
