% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.data.split}
\alias{get.data.split}
\title{get.data.split}
\usage{
get.data.split(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
List of the data split
}
\description{
Function to access the data split info in siamcat@data_split
}
\details{
Access the data split in siamcat@data_split
}
\examples{
 data(siamcat_example)
 data_split <- get.data.split(siamcat_example)
}
