% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.eval_data}
\alias{get.eval_data}
\title{get.eval_data}
\usage{
get.eval_data(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
Evaluation data
}
\description{
Function to access evaluation data in siamcat@eval_data
}
\details{
Access evaluation data in siamcat@eval_data
}
\examples{
 data(siamcat_example)
 eval_data <- get.eval_data(siamcat_example)
}
