% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.features.matrix}
\alias{get.features.matrix}
\title{get.features.matrix}
\usage{
get.features.matrix(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
Features as a matrix
}
\description{
Function to access features in siamcat@phylose@otu_table
}
\details{
Access features in siamcat@phylose@otu_table as matrix
}
\examples{
    data(siamcat_example)
    feat <- get.features.matrix(siamcat_example)
}
