% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.features}
\alias{get.features}
\title{get.features}
\usage{
get.features(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
Features as \link[phyloseq]{otu_table-class} object
}
\description{
Function to access features in siamcat@phylose@otu_table
}
\details{
Access features in siamcat@phylose@otu_table
}
\examples{
 data(siamcat_example)
 feat <- get.features(siamcat_example)
}
