% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.filt_feat.matrix}
\alias{get.filt_feat.matrix}
\title{Retrieve the filtered features from a SIAMCAT object}
\usage{
get.filt_feat.matrix(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
containing filtered features}
}
\value{
A matrix containing the filtered features
}
\description{
Function to retrieve the filtered features from a SIAMCAT
object
}
\details{
The function returns the filtered features as matrix.
See \link{filter.features} for more details.
}
\examples{
data(siamcat_example)
feat.filt <- get.filt_feat.matrix(siamcat_example)
feat.filt[1:3, 1:3]
}
