% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.label.info}
\alias{get.label.info}
\title{get.label.info}
\usage{
get.label.info(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
List of label informations
}
\description{
Function to access label info in siamcat@label@info
}
\details{
Access label info in siamcat@label@info
}
\examples{
 data(siamcat_example)
 label_info <- get.label.info(siamcat_example)
}
