% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.label.label}
\alias{get.label.label}
\title{get.label.label}
\usage{
get.label.label(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
Labels as a vector
}
\description{
Function to access labels in siamcat@label@label
}
\details{
Access labels in siamcat@label@label
}
\examples{
 data(siamcat_example)
 label <- get.label.label(siamcat)
}
