% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.label.list}
\alias{get.label.list}
\title{get.label.list}
\usage{
get.label.list(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
Label object converted to a list
}
\description{
Function to access labels in siamcat@label
}
\details{
Access labels in siamcat@label
}
\examples{
 data(siamcat_example)
 label <- get.label.list(siamcat_example)
}
