% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.label}
\alias{get.label}
\alias{get.label}
\alias{get.label.list}
\alias{get.label}
\alias{get.label.info}
\title{get.label}
\usage{
get.label(siamcat)

get.label.list(siamcat)

get.label.info(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}

\item{siamcat}{an object of class \link{siamcat-class}}

\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
Labels as a vector

Label object converted to a list

List of label informations
}
\description{
Function to access labels in siamcat@label@label

Function to access labels in siamcat@label

Function to access label info in siamcat@label@info
}
\details{
Access labels in siamcat@label@label

Access labels in siamcat@label

Access label info in siamcat@label@info
}
\examples{
 data(siamcat_example)
 label <- get.label(siamcat)
 data(siamcat_example)
 label <- get.label.list(siamcat)
 data(siamcat_example)
 label_info <- get.label.info(siamcat)
}
