% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.model.type}
\alias{get.model.type}
\title{get.model.type}
\usage{
get.model.type(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
Character string specifing the name of the method used to construct the model.
}
\description{
Function to access features in siamcat@model_list@model.type
}
\details{
Access model type in siamcat@model_list@model.type
}
\examples{
 data(siamcat_example)
 get.model.type(siamcat)
}
