% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.model_list}
\alias{get.model_list}
\title{get.model_list}
\usage{
get.model_list(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
List of models in a \link{model_list-class} object
}
\description{
Function to access model_list in siamcat@get.model_list
}
\details{
Access features in siamcat@phylose@otu_table
}
\examples{
 data(siamcat_example)
 model_list <- get.model_list(siamcat_example)
}
