% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.models}
\alias{get.models}
\title{get.models}
\usage{
get.models(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
List of models
}
\description{
Function to access models in siamcat@model_list@models
}
\details{
Access models in siamcat@model_list@models
}
\examples{
 data(siamcat_example)
 models <- get.models(siamcat)
}
