% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.norm_feat.matrix}
\alias{get.norm_feat.matrix}
\title{Retrieve the normalized features from a SIAMCAT object}
\usage{
get.norm_feat.matrix(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
containing normalized features}
}
\value{
A matrix containing the normalized features
}
\description{
Function to retrieve the normalized features from a SIAMCAT
object
}
\details{
The function returns the normalized features as matrix.
See \link{normalize.features} for more details.
}
\examples{
data(siamcat_example)
feat.norm <- get.norm_feat.matrix(siamcat_example)
feat.norm[1:3, 1:3]
}
