% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.orig_feat.matrix}
\alias{get.orig_feat.matrix}
\title{Retrieve the original features from a SIAMCAT object}
\usage{
get.orig_feat.matrix(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}}
}
\value{
A matrix containing the original features
}
\description{
Function to retrieve the original features from a SIAMCAT
object
}
\details{
The function returns the original features as matrix.
}
\examples{
data(siamcat_example)
feat.original <- get.orig_feat.matrix(siamcat_example)
feat.original[1:3, 1:3]
}
