% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.phyloseq}
\alias{get.phyloseq}
\title{get.phyloseq}
\usage{
get.phyloseq(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
Object of class \link[phyloseq]{phyloseq-class}
}
\description{
Function to access phyloseq object in siamcat@phylose
}
\details{
Access phyloseq object in siamcat@phyloseq
}
\examples{
 data(siamcat_example)
 phyloseq <- get.phyloseq(siamcat_example)
}
