% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.pred_matrix}
\alias{get.pred_matrix}
\title{get.pred_matrix}
\usage{
get.pred_matrix(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
Prediction matrix
}
\description{
Function to access prediction matrix  in siamcat@pred_matrix 
siamcat@orig_feat in an object of class \link{siamcat-class}
}
\details{
Access Prediction matrix  in siamcat@pred_matrix
}
\examples{
 data(siamcat_example)
 pred_matrix <- get.pred_matrix(siamcat_example)
}
