% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_plot.R
\name{get.weights.matrix}
\alias{get.weights.matrix}
\title{function to get weights matrix from the model list}
\usage{
get.weights.matrix(models.list, verbose = 0)
}
\description{
function to get weights matrix from the model list
}
\keyword{internal}
