% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{label-class}
\alias{label-class}
\title{The S4 class for storing label info.}
\description{
The S4 class for storing label info.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{inherited from \code{\link{list}} class, contains
a list with:
\itemize{
    \item \code{label} numeric vector, specifying to which category samples
    belong, usualy 1 and -1
    \item \code{type} contains information about the label type
    \item \code{info} list with additional informations about the dataset
}}
}}

