% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ordination.R
\name{make.ordination}
\alias{make.ordination}
\title{Generate ordination of a SIAMCAT object}
\usage{
make.ordination(siamcat, distance="bray", method="Pcoa")
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{distance}{string, distance metric passed to
\code{\link[phyloseq]{ordinate}}, defaults to \code{"bray"}
(Bray-Curtis dissimilarity)}

\item{method}{string, ordination method passed to
\code{\link[phyloseq]{ordinate}}. Supported methods include
\code{c("PCoA", "MDS", "NMDS", "DPCoA", "CAP", "RDA", "CCA", "DCA")},
defaults to \code{"PCoA"}}
}
\description{
A thin wrapper around 
\code{\link[phyloseq]{ordinate}}.
Computes and stores an ordination of the
phyloseq object contained within a \link{siamcat-class} object
in the ordination(siamcat) slot.
}
