% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{make.predictions}
\alias{make.predictions}
\title{Prediction on the test set}
\usage{
make.predictions(siamcat)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}
}
\value{
object of class \link{siamcat-class}
}
\description{
This function takes the test set instances and the model trained by \link{plm.trainer} in order to predict the classes.
}
\keyword{SIAMCAT}
\keyword{plm.predictor}
