% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_evaler_options}
\alias{make_evaler_options}
\title{This is a helper function to uniformly pass command line parameters for model evaluation to SIAMCAT (as well as potentially for the testing framework)}
\usage{
make_evaler_options()
}
\description{
This is a helper function to uniformly pass command line parameters for model evaluation to SIAMCAT (as well as potentially for the testing framework)
}
\examples{
make_evaler_options()
}
