% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_filter_options}
\alias{make_filter_options}
\title{This is a helper function to uniformly pass command line parameters forgeneric  filtering to SIAMCAT as well as the testing framework.}
\usage{
make_filter_options()
}
\description{
This is a helper function to uniformly pass command line parameters forgeneric  filtering to SIAMCAT as well as the testing framework.
}
\examples{
make_filter_options()
}
