% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_frozen_normalizer_options}
\alias{make_frozen_normalizer_options}
\title{This is a helper function to uniformly pass command line parameters for frozen normalization to SIAMCAT}
\usage{
make_frozen_normalizer_options()
}
\description{
This is a helper function to uniformly pass command line parameters for frozen normalization to SIAMCAT
}
\examples{
make_frozen_normalizer_options()
}
