% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_meta_adder_options}
\alias{make_meta_adder_options}
\title{This is a helper function to uniformly pass command line parameters for metadata addition to SIAMCAT}
\usage{
make_meta_adder_options()
}
\description{
This is a helper function to uniformly pass command line parameters for metadata addition to SIAMCAT
}
\examples{
make_meta_adder_options()
}
\keyword{SIAMCAT}
\keyword{make_evaler_options}
