% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_model_interpretor_options}
\alias{make_model_interpretor_options}
\title{This is a helper function to uniformly pass command line parameters for model interpretation in SIAMCAT}
\usage{
make_model_interpretor_options()
}
\description{
This is a helper function to uniformly pass command line parameters for model interpretation in SIAMCAT
}
\examples{
make_model_interpretor_options()
}
\keyword{SIAMCAT}
\keyword{make_evaler_options}
