% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{make_normalizer_options}
\alias{make_normalizer_options}
\title{This is a helper function to uniformly pass command line parameters for generic normalizing to SIAMCAT as well as the testing framework.}
\usage{
make_normalizer_options()
}
\description{
This is a helper function to uniformly pass command line parameters for generic normalizing to SIAMCAT as well as the testing framework.
}
\examples{
make_normalizater_options()
}
