% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{meta.crc.zeller}
\alias{meta.crc.zeller}
\title{Documentation for the example metadata object in the data folder}
\description{
Metadata (as data.frame) for the CRC dataset, containing 6 variables (e.g.
Age or BMI) for 141 samples.
}
\keyword{data}
