% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{model_list-class}
\alias{model_list-class}
\title{The S4 class for storing models.}
\description{
The S4 class for storing models.
}
\section{Slots}{

\describe{
\item{\code{models}}{a list with models obtained from \link{train.model}}

\item{\code{model.type}}{name of the method used by \link{train.model}}

\item{\code{feature.type}}{which types of features used by \link{train.model}}
}}

