% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{models}
\alias{models}
\alias{models,ANY-method}
\title{Retrieve list of models from object.}
\usage{
models(siamcat, verbose=1)

\S4method{models}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a model_list or instance of \link{model_list-class}.}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The list of models or NULL.
}
\description{
Retrieve list of models from object.
}
\examples{
    data(siamcat_example)
    models(siamcat_example)
}
