% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{norm_feat-class}
\alias{norm_feat-class}
\title{The S4 class for storing the normalization data/parameters}
\description{
The S4 class for storing the normalization data/parameters
}
\section{Slots}{

\describe{
\item{\code{norm.feat}}{An object of class \link[phyloseq]{otu_table-class} storing
the normalized features}

\item{\code{norm.param}}{A list with:
    \itemize{
    \item \code{norm.method} the normalization method used
    \item \code{retained.feat} the names of features retained after filtering
    \item \code{log.n0} pseudocount
    \item \code{n.p} vector norm
    \item \code{norm.margin} margin for the normalization
} and additional entries depending on the normalization method used.}
}}

