% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{norm_feat}
\alias{norm_feat}
\alias{norm_feat,ANY-method}
\title{Retrieve normalized features form object}
\usage{
norm_feat(siamcat, verbose=1)

\S4method{norm_feat}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains an object in the norm_feat slot}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The normalized feature matrix or NULL.
}
\description{
Retrieve normalized features form object
}
\examples{
    data(siamcat_example)
    norm_feat(siamcat_example)
}
