% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{norm_params}
\alias{norm_params}
\alias{norm_params,ANY-method}
\title{Retrieve the list of normalization parameters from object.}
\usage{
norm_params(siamcat, verbose=1)

\S4method{norm_params}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a norm_feat}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The list of normalization parameters or NULL.
}
\description{
Retrieve the list of normalization parameters from object.
}
\examples{
    data(siamcat_example)
    norm_params(siamcat_example)
}
