% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{normalize.feat}
\alias{normalize.feat}
\title{This function normalizes features according to specified parameters.}
\usage{
normalize.feat(feat, norm.method, log.n0, sd.min.q, n.p, n.sample, n.feature,
  n.global)
}
\arguments{
\item{norm.method}{Normalization method}

\item{log.n0}{Pseudocount to be added before log-transformation}

\item{sd.min.q}{Quantile of the distribution of standard deviation of all feature that will be added to the denominator during standardization of each feature in order to avoid underestimation (only for metod==\"log.std\")'}

\item{n.p}{Vector norm to use}

\item{n.sample}{Normalize by feature?}

\item{n.feature}{Normalize by sample?}

\item{n.global}{Normalize by global rescaling?}

\item{fn.in.feat}{feature input file}

\item{fn.out.feat}{feature output file}

\item{fn.out.param}{output file for normalization parameters}
}
\description{
This function normalizes features according to specified parameters.
}
\examples{
normalize.feat()
}
\keyword{SIAMCAT}
\keyword{association_check}
