% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{ordination}
\alias{ordination}
\title{Retrieve ordination}
\usage{
ordination(siamcat)
}
\arguments{
\item{siamcat}{(Required). A \link{siamcat-class} object
that contains evaluation data}
}
\value{
A list with ordination ("ord") and parameters ("method", "distance").
}
\description{
Function to retrieve the SIAMCAT ordination for the
contained phyloseq object as generated by make.ordination.
}
\examples{
ordination(siamcat_example)
}
