% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{orig_feat-class}
\alias{orig_feat-class}
\title{The S4 class for storing original features info.}
\description{
The S4 class for storing original features info.
}
\section{Slots}{

\describe{
\item{\code{taxa_are_rows}}{A single logical specifying the orientation
of the abundance table}

\item{\code{.Data}}{inherited from \code{\link{matrix}} class, contains
a matrix with predictions made by \link{make.predictions} function}
}}

