% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labels.r
\name{parse.label.header}
\alias{parse.label.header}
\title{Parse label header}
\usage{
parse.label.header(label.header)
}
\arguments{
\item{label.header}{- string in the format: #<TYPE>:<L1>=<class1>;
<L2>=<class2>[;<L3>=<class3>] where <TYPE> is a string specifying the type
of label variable such as BINARY (for binary classification), CATEGORICAL
(for multi-class classification), or CONTINUOUS (for regression)
<L1> is a short numeric label for the first class with description <class1>
(similarly for the other classes)}
}
\value{
a list with tow items \itemize{
\item \code{$type} type of the label: BINARY CONTINUOUS or CATEGORICAL
\item \code{$class.descr} lables and information on what do they mean
}
}
\description{
This function parses the header of a label file
}
\keyword{internal}
