% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_predictor.r
\name{plm.predictor}
\alias{plm.predictor}
\title{Prediction on the test set}
\usage{
plm.predictor(feat, label, data.split = NULL, models.list, model.mat,
  model.type)
}
\arguments{
\item{feat}{features object}

\item{label}{label object}

\item{data.split}{filename containing the test samples or list of test instances produced by \link{data.splitter()}, defaults to \code{NULL} leading to testing on the complete dataset}

\item{model.mat}{model matrix needed to rebuild the model}

\item{model.type}{string, type of the model that was trained}

\item{model}{model object trained by \link{plm.trainer}}
}
\value{
list containing the precitions \itemize{
 \item \code{$pred};
 \item \code{$mat}
}
}
\description{
This function takes the test set instances and the model trained by \link{plm.trainer} in order to predict the classes.
}
\keyword{SIAMCAT}
\keyword{plm.predictor}
