% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{plm.trainer}
\alias{plm.trainer}
\title{Model training}
\usage{
plm.trainer(feat, label, fn.train.sample, num.folds = 5, stratify,
  modsel.crit, min.nonzero.coeff)
}
\arguments{
\item{feat}{features object}

\item{label}{label object}

\item{fn.train.sample}{file containing the training samples}

\item{num.folds}{integer, number of folds for the internal model cross-validation}

\item{stratify}{boolean, should the folds in the internal cross-validation be stratified?}

\item{modsel.crit}{model selection criterion (not used at the moment)}

\item{min.nonzero.coeff}{integer number of minimum nonzero coefficients that should be present in the model}
}
\value{
list containing \itemize{
\item \code{$out.matrix};
\item \code{$model.header};
\item \code{$W.mat};
\item \code{$hyperpar.mat};
\item \code{$model}
}
}
\description{
This function trains the a machine learning model on the training data, using a \code{num.folds}-fold internal cross-validation scheme to find the optimal hyper-parameters of the model.
}
\keyword{SIAMCAT}
\keyword{plm.trainer}
