% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_plot.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{function to plot the heatmap}
\usage{
\method{plot}{heatmap}(image.data, limits, color.scheme, effect.size,
  verbose = 0)
}
\description{
function to plot the heatmap
}
\keyword{internal}
