% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{pred_matrix-class}
\alias{pred_matrix-class}
\title{The S4 class for storing predictions.}
\description{
The S4 class for storing predictions.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{inherited from \code{\link{matrix}} class, contains
a matrix with predictions made by \link{make.predictions} function}
}}

