% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_plot.R
\name{prepare.heatmap.zscore}
\alias{prepare.heatmap.zscore}
\title{function to prepare the data to plot zscore heatmap}
\usage{
prepare.heatmap.zscore(heatmap.data, limits, verbose = 0)
}
\description{
function to prepare the data to plot zscore heatmap
}
\keyword{internal}
