% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_utils.r
\name{read.features}
\alias{read.features}
\title{Read feature file}
\usage{
read.features(fn.in.feat)
}
\arguments{
\item{fn.in.feat}{name of the tsv file containing features}
}
\value{
matrix containing features from the file
}
\description{
This file reads in the tsv file with features and converts it into a matrix.

The file should be oragnized as follows:
features (in rows) x samples (in columns).

First row should contain sample labels (consistent with label data), while the first column should contain feature labels (e.g. taxonomic identifiers). The remaining entries are expected to be real values \code{>= 0} that quantify the abundance of each feature in each sample.
}
