% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labels.r
\name{read.labels}
\alias{read.labels}
\title{Read labels file}
\usage{
read.labels(fn.in.label)
}
\arguments{
\item{fn.in.label}{name of the tsv file containing labels}
}
\value{
label object containing several entries:\itemize{
\item \code{$label} named vector containing the numerical labels from the
file;
\item \code{$info} information about the classes in the label;
\item \code{$type} information about the label type (e.g. \code{BINARY});
}
}
\description{
This file reads in the tsv file with labels and converts it
into a label object.

First row is expected to be \code{#BINARY:1=[label for cases];
-1=[label for controls]}.
Second row should contain the sample identifiers as tab-separated list
(consistent with feature and metadata).

Third row is expected to contain the actual class labels (tab-separated):
\code{1} for each case and \code{-1} for each control.

Note: Labels can take other numeric values (but not characters or strings);
importantly, the label for cases has to be greater than the one for controls
}
\examples{
    # run with example data
fn.label <- system.file('extdata', 'label_crc_zeller_msb_mocat_specI.tsv',
    package = 'SIAMCAT')

labels <- read.labels(fn.label)
}
