% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{read.meta}
\alias{read.meta}
\title{Read metadata file}
\usage{
read.meta(fn.in.meta)
}
\arguments{
\item{fn.in.meta}{name of the tsv file containing metadata}
}
\value{
matrix containing features from the file
}
\description{
This file reads in the tsv file with metadata and converts it into a matrix
}
