% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{reset.features}
\alias{reset.features}
\title{reset.features}
\usage{
reset.features(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
A new \link{siamcat-class} object
}
\description{
Function reset features in siamcat@phylose@otu_table to those in
siamcat@orig_feat in an object of class \link{siamcat-class}
}
\details{
Reset features in siamcat@phylose@otu_table to those in siamcat@orig_feat
}
\examples{
 data(siamcat_example)
 siamcat_example <- reset.features(siamcat_example)
}
