% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interpretation_plot.R
\name{select.features}
\alias{select.features}
\title{function to select the features to plot on the heatmap}
\usage{
select.features(weights, model.type, consens.thres, norm.models, label,
  max.show, verbose = 0)
}
\description{
function to select the features to plot on the heatmap
}
\keyword{internal}
