% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class.R
\docType{class}
\name{siamcat-class}
\alias{siamcat-class}
\title{The S4 class for storing taxa-abundance information and models.}
\description{
The S4 class for storing taxa-abundance information and models.
}
