% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefseformat.R
\name{siamcat.to.lefse}
\alias{siamcat.to.lefse}
\title{create a lefse input file from siamcat object}
\usage{
siamcat.to.lefse(siamcat, filename = "siamcat_output.txt")
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{filename}{name of the input file to which data will be save}
}
\value{
nothing but data is written to a file
}
\description{
This function creates a lefse input file from siamcat object
}
\examples{

data(siamcat_example)
siamcat.to.lefse(siamcat_example)

}
\keyword{siamcat.to.lefse}
