% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maaslinformat.R
\name{siamcat.to.maaslin}
\alias{siamcat.to.maaslin}
\title{create a MaAsLin input file from siamcat object}
\usage{
siamcat.to.maaslin(siamcat, filename = "siamcat_output.pcl")
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{filename}{name of the input file to which data will be save}
}
\value{
nothing but data is written to a file
}
\description{
This function creates a MaAsLin merged PCL single input file 
from siamcat object
}
\examples{

data(siamcat_example)
siamcat.to.maaslin(siamcat_example)

}
\keyword{internal}
