% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class.R
\docType{data}
\name{siamcat_example}
\alias{siamcat_example}
\title{Documentation for the example siamcat object in the data folder}
\description{
Documentation for the example siamcat object in the data folder
}
\keyword{data}
