% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{siamcat_simple}
\alias{siamcat_simple}
\title{Documentation for the example siamcat object in the data folder}
\description{
Documentation for the example siamcat object in the data folder
}
\keyword{data}
