% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.r
\name{split.data}
\alias{split.data}
\title{Split a dataset into training and a test sets.}
\usage{
\method{split}{data}(siamcat, num.folds = 2, num.resample = 1,
  stratify = TRUE, inseparable = NULL, verbose = 1)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{num.folds}{number of cross-validation folds (needs to be \code{>=2}), defaults to \code{2}}

\item{num.resample}{resampling rounds (values \code{<= 1} deactivate resampling), defaults to \code{1}}

\item{stratify}{boolean, should the splits be stratified s. t. an equal proportion of classes are present in each fold?, defaults to \code{TRUE}}

\item{inseparable}{column index or column name of metadata variable, defaults to \code{NULL}}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for only information about progress and success, \code{2} for normal 
level of information and \code{3} for full debug information, defaults to \code{1}}
}
\value{
object of class \link{siamcat-class}
}
\description{
This function prepares the cross-validation by splitting the data into \code{num.folds} training and test folds for \code{num.resample} times.
}
\keyword{SIAMCAT}
\keyword{split.data}
