% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_utils.r
\name{train.plm}
\alias{train.plm}
\title{Perform feature normalization}
\usage{
train.plm(data, method = c("lasso", "enet", "ridge", "lasso_ll", "ridge_ll",
  "randomForest"), measure = list("acc"), min.nonzero.coeff = 5,
  param.set = NULL, neg.lab, verbose = 1)
}
\arguments{
\item{data}{a dataframe created from features and lable as last column}
}
\description{
This function performs feature normalization according to user-
 specified parameters.
}
\keyword{internal}
