% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.r
\name{validate.data}
\alias{validate.data}
\title{Validate samples in metadata, labels, and features}
\usage{
validate.data(feat, label, meta = NULL)
}
\arguments{
\item{feat}{features object}

\item{label}{labels object}

\item{meta}{metadata object}
}
\value{
a list containing values after validation: \itemize{
 \item \code{$feat} = features;
 \item \code{$label} = labels;
 \item \code{$meta} = metadata
}
}
\description{
This function checks if labels are available for all samples in features. Additionally validates metadata, if given
}
\keyword{SIAMCAT}
\keyword{validate.data}
