% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{versions}
\alias{versions}
\title{Retrieve version numbers}
\usage{
versions(siamcat)
}
\arguments{
\item{siamcat}{(Required). A \link{siamcat-class} object
that contains evaluation data}
}
\value{
A list with package names and version numbers
}
\description{
Function to retrieve the SIAMCAT and phyloseq version number
from a SIAMCAT object
}
\details{
The functions returns a list containing the package version numbers
from a SIAMCAT object.
}
\examples{
versions(siamcat_example)
}
