% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{accessSlot}
\alias{accessSlot}
\title{Universal slot accessor function for siamcat-class.}
\usage{
accessSlot(siamcat, slot, errorIfNULL=FALSE)
}
\arguments{
\item{siamcat}{an object of \link{siamcat-class}.}

\item{slot}{A character string indicating the slot (not data class)
of the component data type that is desired.}
}
\value{
Returns the component object specified by the argument \code{slot}. 
 Returns NULL if slot does not exist.
}
\description{
This function is used internally by many accessors.
}
\examples{
#
data(siamcat_example)
access(siamcat_example, "label")
access(siamcat_example, "model_list")
}
