% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta_pred.r
\name{add.meta.pred}
\alias{add.meta.pred}
\title{Add metadata as predictors}
\usage{
add.meta.pred(feat, meta, pred.names = NULL, std.meta)
}
\arguments{
\item{feat}{features object}

\item{meta}{metadata object}

\item{pred.names}{vector of names of the metavariables to be added to the feature matrix as predictors}

\item{std.meta}{boolean, should added metadata features be standardized?}
}
\value{
features object with added metadata
}
\description{
This function adds metadata to the feature matrix to be later used as predictors
}
\keyword{SIAMCAT}
\keyword{add.meta.pred}
