% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{data_split<-}
\alias{data_split<-}
\alias{assign-data_split}
\alias{data_split<-,siamcat,data_split-method}
\alias{data_split<-}
\title{Assign a new data_split object to \code{x}}
\usage{
data_split(x) <- value

\S4method{data_split}{siamcat,data_split}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link{data_split-class}}
}
\value{
none
}
\description{
Assign a new data_split object to \code{x}
}
\examples{
data(siamcat_example)
data_split(siamcat_example) <- data_split(siamcat_example)
}
