% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{eval_data<-}
\alias{eval_data<-}
\alias{assign-eval_data}
\alias{eval_data<-,siamcat,list-method}
\alias{eval_data<-}
\title{Assign a new eval_data object to \code{x}}
\usage{
eval_data(x) <- value

\S4method{eval_data}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an eval_data list}
}
\value{
none
}
\description{
Assign a new eval_data object to \code{x}
}
\examples{
data(siamcat_example)
eval_data(siamcat_example) <- eval_data(siamcat_example)
}
