% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{label<-}
\alias{label<-}
\alias{assign-label}
\alias{label<-,siamcat,label-method}
\alias{label<-}
\title{Assign a new label object to \code{x}}
\usage{
label(x) <- value

\S4method{label}{siamcat,label}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link{label-class}}
}
\value{
none
}
\description{
Assign a new label object to \code{x}
}
\examples{
data(siamcat_example)
label(siamcat_example) <- label(siamcat_example)
}
