% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{model_list<-}
\alias{model_list<-}
\alias{assign-model_list}
\alias{model_list<-,siamcat,model_list-method}
\alias{model_list<-}
\title{Assign a new model_list object to \code{x}}
\usage{
model_list(x) <- value

\S4method{model_list}{siamcat,model_list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link{model_list-class}}
}
\value{
none
}
\description{
Assign a new model_list object to \code{x}
}
\examples{
data(siamcat_example)
model_list(siamcat_example) <- model_list(siamcat_example)
}
