% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{norm_feat<-}
\alias{norm_feat<-}
\alias{assign-norm_feat}
\alias{norm_feat<-,siamcat,list-method}
\title{Assign a new list containing normalziation parameters and normalized
features to a SIAMCAT object}
\usage{
norm_feat(x) <- value

\S4method{norm_feat}{siamcat,list}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{a list containing normaliation parameters and features}
}
\value{
none
}
\description{
Assign a new list containing normalziation parameters and normalized
features to a SIAMCAT object
}
\examples{
data(siamcat_example)
norm_feat(siamcat_example) <- norm_feat(siamcat_example)
}
\keyword{internal}
