% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_assignment_methods.R
\docType{methods}
\name{orig_feat<-}
\alias{orig_feat<-}
\alias{assign-orig_feat}
\alias{orig_feat<-,siamcat,orig_feat-method}
\alias{orig_feat<-}
\alias{orig_feat<-,siamcat,otu_table-method}
\alias{orig_feat<-}
\title{Assign a new otu_table object to \code{x} orig_feat slot}
\usage{
orig_feat(x) <- value

\S4method{orig_feat}{siamcat,orig_feat}(x) <- value

\S4method{orig_feat}{siamcat,otu_table}(x) <- value
}
\arguments{
\item{x}{an object of class \link{siamcat-class}}

\item{value}{an object of class \link[phyloseq]{otu_table-class}}
}
\value{
none
}
\description{
Assign a new otu_table object to \code{x} orig_feat slot
}
\examples{
data(siamcat_example)
orig_feat(siamcat_example) <- orig_feat(siamcat_example)
}
