% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_split.r
\name{assign.fold}
\alias{assign.fold}
\title{Assign folds}
\usage{
assign.fold(label, num.folds, stratified, inseparable = NULL, meta = NULL,
  verbose = 1)
}
\value{
vector with fold assignments
}
\description{
Create data split from label
}
\keyword{internal}
