% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{associations}
\alias{associations}
\title{Retrieve the results of association testing from a SIAMCAT object}
\usage{
associations(siamcat, verbose=1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
containing the results of association testing}

\item{verbose}{integer, if the slot is empty, should a message be printed?
values can be either \code{0} (no output) or \code{1} (print message)}
}
\value{
A \code{data.frame} of association testing results or \code{NULL}
}
\description{
Function to retrieve the results of association testing
}
\details{
The function returns the results of the association testing
procedure as dataframe. See \link{check.associations} for more details.
}
\examples{
data(siamcat_example)
temp <- associations(siamcat_example)
head(temp)
}
