% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{associations}
\alias{associations}
\alias{associations,ANY-method}
\title{Retrieve associations from object.}
\usage{
associations(siamcat, verbose=1)

\S4method{associations}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains an object in the associations slot}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The results of the association testing or NULL.
}
\description{
Retrieve associations from object.
}
\examples{
    data(siamcat_example)
    associations(siamcat_example)
}
