% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_associations.r
\name{check.associations}
\alias{check.associations}
\title{Check and visualize associations between features and classes}
\usage{
check.associations(siamcat, fn.plot, color.scheme = "RdYlBu", alpha = 0.05,
  mult.corr = "fdr", sort.by = "fc", detect.lim = 10^-8,
  pr.cutoff = 10^-6, max.show = 50, plot.type = "quantile.box",
  panels = c("fc", "auroc"))
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{fn.plot}{filename for the pdf-plot}

\item{color.scheme}{valid R color scheme, defaults to \code{'RdYlBu'}}

\item{alpha}{float, significance level, defaults to \code{0.05}}

\item{mult.corr}{multiple hypothesis correction method, see \code{\link[stats]{p.adjust}}, defaults to \code{"fdr"}}

\item{sort.by}{string, sort features by p-value (\code{"pv"}), by fold change (\code{"fc"}) or by prevalence shift (\code{"pr.shift"}), defaults to \code{"pv"}}

\item{detect.lim}{float, pseudocount to be added before log-transormation of the data, defaults to \code{1e-08}}

\item{pr.cutoff}{float, cutoff for the prevalence computation, defaults to \code{1e-06}}

\item{max.show}{integer, how many associated features should be shown, defaults to \code{50}}

\item{plot.type}{string, specify how the abundance should be plotted, must be one of these: \code{c("bean", "box", "quantile.box", "quantile.rect")}, defaults to \code{"quantile.box"}}

\item{panels}{vector, name of the panels to be plotted next to the log10-transformed abundances, possible entries are \code{c("fc", "auroc", "prevalence")}, defaults to \code{c("fc", "auroc")}}
}
\value{
Does not return anything, but produces an association plot
}
\description{
This function calculates for each feature a pseudo-fold change
(geometrical mean of the difference between quantiles)
between the different classes found in labels.

Significance of the differences is computed for each feature using a wilcoxon
test followed by multiple hypothesis testing correction.

Additionally, the Area under the Receiver Operating Characteristic Curve
(AU-ROC) and a prevalence shift are computed for the features found to be
associated with the two different classes at a user-specified
significance level \code{alpha}.

Finally, the function produces a plot of the top \code{max.show} associated
features, showing the distribution of the log10-transformed abundances for
both classes, and user-selected panels for the effect (AU-ROC, Prevalence
Shift, and Fold Change)
}
\keyword{SIAMCAT}
\keyword{check.associations}
