% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_associations.r
\name{check.associations}
\alias{check.associations}
\title{Check and visualize associations between features and classes}
\usage{
check.associations(siamcat, fn.plot=NULL, color.scheme = "RdYlBu",
    alpha =0.05, mult.corr = "fdr", sort.by = "fc",
    detect.lim = 1e-06, pr.cutoff = 1e-6, max.show = 50,
    plot.type = "quantile.box",
    panels = c("fc","auroc"), prompt = TRUE,
    feature.type = 'filtered', verbose = 1)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{fn.plot}{string, filename for the pdf-plot. If \code{fn.plot} is
\code{NULL}, the plot will be produced in the active graphics device.}

\item{color.scheme}{valid R color scheme or vector of valid R colors (must be
of the same length as the number of classes), defaults to \code{'RdYlBu'}}

\item{alpha}{float, significance level, defaults to \code{0.05}}

\item{mult.corr}{string, multiple hypothesis correction method, see
\code{\link[stats]{p.adjust}}, defaults to \code{"fdr"}}

\item{sort.by}{string, sort features by p-value (\code{"p.val"}), by fold
change (\code{"fc"}) or by prevalence shift (\code{"pr.shift"}),
defaults to \code{"fc"}}

\item{detect.lim}{float, pseudocount to be added before log-transformation of
the data, defaults to \code{1e-06}. Will be ignored if
\code{feature.type} is \code{"normalized"}.}

\item{pr.cutoff}{float, cutoff for the prevalence computation, defaults to
\code{1e-06}}

\item{max.show}{integer, how many associated features should be shown,
defaults to \code{50}}

\item{plot.type}{string, specify how the abundance should be plotted, must be
one of these: \code{c("bean", "box", "quantile.box", "quantile.rect")},
defaults to \code{"quantile.box"}}

\item{panels}{vector, name of the panels to be plotted next to the
abundances, possible entries are \code{c("fc", "auroc",
"prevalence")}, defaults to \code{c("fc", "auroc")}}

\item{prompt}{boolean, turn on/off prompting user input when not plotting
into a pdf-file, defaults to TRUE}

\item{feature.type}{string, on which type of features should the function
work? Can be either \code{c()"original", "filtered", or "normalized")}.
Please only change this paramter if you know what you are doing!

If \code{feature.type} is \code{"normalized"}, the normalized abundances
will not be log10-transformed.}

\item{verbose}{integer, control output: \code{0} for no output at all,
\code{1} for only information about progress and success, \code{2} for
normal level of information and \code{3} for full debug information,
defaults to \code{1}}
}
\value{
object of class \link{siamcat-class} with the slot
    \code{associations} filled
}
\description{
This function computes different measures of association between
    features and the label and visualizes the results
}
\details{
For each feature, this function calculates different measures of
    association between the feature and the label. In detail, these
    associations are: \itemize{
    \item Significance as computed by a Wilcoxon test followed by multiple
    hypothesis testing correction.
    \item AUROC (Area Under the Receiver Operating Characteristics Curve)
    as a non-parameteric measure of enrichment (corresponds to the effect
    size of the Wilcoxon test).
    \item The generalized Fold Change (gFC) is a pseudo fold change
    which is calculated as geometric mean of the differences between the
    quantiles for the different classes found in the label.
    \item The prevalence shift between the two different classes found in
    the label.
    }

    Finally, the function produces a plot of the top \code{max.show}
    associated features at a user-specified significance level \code{alpha},
    showing the distribution of the log10-transformed abundances for both
    classes, and user-selected panels for the effect (AU-ROC, Prevalence
    Shift, and Fold Change).
}
\examples{
# Example data
data(siamcat_example)

# Simple example
siamcat_example <- check.associations(siamcat_example,
    fn.plot='./assoc_plot.pdf')

# Plot associations as box plot
siamcat_example <- check.associations(siamcat_example,
    fn.plot='./assoc_plot_box.pdf', plot.type='box')

# Additionally, sort by p-value instead of by fold change
siamcat_example <- check.associations(siamcat_example,
    fn.plot='./assoc_plot_fc.pdf', plot.type='box', sort.by='p.val')

# Custom colors
siamcat_example <- check.associations(siamcat_example,
    fn.plot='./assoc_plot_blue_yellow.pdf', plot.type='box',
    color.scheme=c('cornflowerblue', '#ffc125'))
}
\keyword{SIAMCAT}
\keyword{check.associations}
