% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_confounders.r
\name{check.confounders}
\alias{check.confounders}
\title{Check for potential confounders in the metadata}
\usage{
check.confounders(siamcat, fn.plot, verbose = 1)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}

\item{fn.plot}{string, filename for the pdf-plot}

\item{verbose}{control output: \code{0} for no output at all, \code{1} for
only information about progress and success, \code{2} for normal level of
information and \code{3} for full debug information, defaults to \code{1}}
}
\value{
Does not return anything, but produces a single plot for each
    metadata category.
}
\description{
This function checks for associations between class labels and
    potential confounders (e.g. age, sex, or BMI) that are present in the
    metadata. Statistical testing is performed with Fisher's exact test or
    Wilcoxon test, while associations are visualized either as barplot or Q-Q
    plot, depending on the type of metadata.
}
\examples{
    # Example data
    data(siamcat_example)

    # Simple working example
    check.confounders(siamcat_example, './conf_plot.pdf')
}
\keyword{SIAMCAT}
\keyword{check.confounders}
