% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{confounder.check}
\alias{confounder.check}
\title{Check for potential confounders in the metadata}
\usage{
confounder.check(meta, label)
}
\arguments{
\item{meta}{metadata object}

\item{label}{labels object}
}
\value{
Does not return anything, but produces a single plot for each metadata category
}
\description{
This function checks for associations between class labels and potential confounders (e.g. age, sex, or BMI) that are present in the metadata. Statistical testing is performed with Fisher's exact test or Wilcoxon test, while associations are visualized either as barplot or Q-Q plot, depending on the type of metadata.
}
\keyword{SIAMCAT}
\keyword{confounder_check}
