% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{confounder.check}
\alias{confounder.check}
\title{This function checks for potential confounders in the metadata}
\usage{
confounder.check(meta, label)
}
\arguments{
\item{source.dir}{SIAMCAT script directory}

\item{fn.in.meta}{meta-data input file}

\item{fn.in.label}{label input file}
}
\description{
This function checks for potential confounders in the metadata
}
\examples{
confounder.check()
}
\keyword{SIAMCAT}
\keyword{confounder_check}
