% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.r
\name{create.label.from.metadata}
\alias{create.label.from.metadata}
\title{create a label object from metadata}
\usage{
create.label.from.metadata(meta, column)
}
\arguments{
\item{meta}{metadata as read by \link{read.meta}
of \link[phyloseq]{sample_data-class}}

\item{column}{name of column that will be used
to create the label}
}
\value{
an object of class \link{label-class}
}
\description{
This function creates a label object from metadata
}
\examples{
   data(siamcat_example)
   label <- create.label.from.metadata(meta(siamcat_example),"gender")

}
\keyword{internal}
