% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{data.splitter}
\alias{data.splitter}
\title{This function splits a dataset into a training and a test set.}
\usage{
data.splitter(label, num.folds, num.resample, stratify, inseparable, meta)
}
\arguments{
\item{label}{label input file}

\item{num.folds}{Number of cross-validation folds (i.e. subsets, needs to be >= 2)}

\item{num.resample}{Resampling rounds (values <= 1 deactivate resampling)}

\item{stratify}{Should cross-validation be stratified such that an approx. equal proportion of positive examples are contained in each subset (only for binary labels)?')}

\item{inseparable}{defaults to NULL}

\item{label.header}{file containing information about the label header}

\item{fn.train.folds}{Output file containing training sets}

\item{fn.test.folds}{Output file containing test sets}

\item{fn.in.meta}{input file for metadata (only of inseperable is specified)}
}
\description{
This function splits a dataset into a training and a test set.
}
\examples{
data.splitter()
}
\keyword{SIAMCAT}
\keyword{association_check}
