% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitter.r
\name{data.splitter}
\alias{data.splitter}
\title{Split a dataset into training and a test sets.}
\usage{
data.splitter(label, num.folds, num.resample, stratify, inseparable, meta)
}
\arguments{
\item{label}{label object}

\item{num.folds}{number of cross-validation folds (needs to be \code{>=2})}

\item{num.resample}{resampling rounds (values \code{<= 1} deactivate resampling)}

\item{stratify}{boolean, should the splits be stratified s. t. an equal proportion of classes are present in each fold?}

\item{inseparable}{defaults to NULL (obsolete)}

\item{meta}{metadata object (obsolete)}
}
\value{
list containing the indices of the training and test folds and the parameters of the splits: \itemize{
 \item \code{$training.folds} nested list, containing for \code{length(num.folds)} the sample names of the \code{length(num.resample)} training folds;
 \item \code{$test.folds} nested list, containing for \code{length(num.folds)} the sample names of the \code{length(num.resample)} test folds;
 \item \code{$num.resample} = number of repeated samplings;
 \item \code{$num.folds} = number of folds
}
}
\description{
This function prepares the cross-validation by splitting the data into \code{num.folds} training and test folds for \code{num.resample} times.
}
\keyword{SIAMCAT}
\keyword{data.splitter}
