% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{data_split}
\alias{data_split}
\alias{data_split,ANY-method}
\alias{data_split,data_split-method}
\alias{data_split,list-method}
\title{Retrieve a \link{data_split-class} object from object.}
\usage{
data_split(siamcat, verbose=1)

\S4method{data_split}{ANY}(siamcat, verbose = 1)

\S4method{data_split}{data_split}(siamcat)

\S4method{data_split}{list}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link{data_split-class} or a list.}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The \link{data_split-class} object or NULL.
}
\description{
Retrieve a \link{data_split-class} object from object.
}
\examples{
    data(siamcat_example)
    data_split(siamcat_example)
}
