% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{eval_data}
\alias{eval_data}
\alias{eval_data,ANY-method}
\alias{eval_data,list-method}
\title{Retrieve eval_data from object.}
\usage{
eval_data(siamcat, verbose=1)

\S4method{eval_data}{ANY}(siamcat, verbose = 1)

\S4method{eval_data}{list}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a eval_data..}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The eval_data list or NULL.
}
\description{
Retrieve eval_data from object.
}
\examples{
    data(siamcat_example)
    eval_data(siamcat_example)
}
