% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_predictions.r
\name{evaluate.predictions}
\alias{evaluate.predictions}
\title{Evaluate prediction results}
\usage{
evaluate.predictions(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for only information about progress and success, \code{2} for normal 
level of information and \code{3} for full debug information, defaults to \code{1}}
}
\value{
object of class \link{siamcat-class}
}
\description{
This function takes the correct labels and predictions for all samples and evaluates the results using the \itemize{
 \item Area Under the Receiver Operating Characteristic (ROC) Curve (AU-ROC)
 \item and the Precision-recall Curve (PR)
}
as metric. Predictions can be supplied either for a single case or as matrix after resampling of the dataset.

Prediction results are usually produced with the function \link{plm.predictor}.
}
\keyword{SIAMCAT}
\keyword{eval.result}
