% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_evaluation.r
\name{evaluation.model.plot}
\alias{evaluation.model.plot}
\title{Model Evaluation Plot}
\usage{
evaluation.model.plot(label, pred, eval.data, fn.plot)
}
\arguments{
\item{label}{a label object}

\item{pred}{matrix containing the model predictions for every CV repetition}

\item{eval.data}{list of evaluation results produced by \link{eval.results}}

\item{fn.plot}{string, filename for the pdf-plot}
}
\value{
Does not return anything, but produces the model evaluation plot.
}
\description{
Produces two plots for model evaluation. The first plot shows the Receiver Operating Receiver (ROC)-curves, the other the Precision-recall (PR)-curves for the different CV repetitions.
}
\keyword{SIAMCAT}
\keyword{evaluation.model.plot}
