% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\encoding{UTF-8}
\name{feat.crc.zeller}
\alias{feat.crc.zeller}
\title{Example feature matrix}
\source{
\url{http://msb.embopress.org/content/10/11/766}
}
\description{
Feature matrix (as data.frame) of the CRC dataset from Zeller
et al. MSB 2014 (see \url{http://msb.embopress.org/content/10/11/766}),
containing 141 samples and 1754 bacterial species (features).
}
\keyword{data}
