% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{feature_type}
\alias{feature_type}
\alias{feature_type,ANY-method}
\title{Retrieve feature_type from object.}
\usage{
feature_type(siamcat)

\S4method{feature_type}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a model_list or instance of \link{model_list-class}.}
}
\value{
The string describing type of feature used for the model or NULL.
}
\description{
Retrieve feature_type from object.
}
\examples{
    data(siamcat_example)
    feature_type(siamcat_example)
}
