% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{feature_weights}
\alias{feature_weights}
\alias{feature_weights,ANY-method}
\title{Retrieve feature_weights from object.}
\usage{
feature_weights(siamcat)

\S4method{feature_weights}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a model_list or instance of \link{model_list-class}.}
}
\value{
A dataframe containing mean/median feature weight and additional info
}
\description{
Retrieve feature_weights from object.
}
\examples{
    data(siamcat_example)
    feature_weights(siamcat_example)
}
