% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,ANY-method}
\alias{features,otu_table-method}
\title{Retrieve a \link[phyloseq]{otu_table-class} object from object.}
\usage{
features(siamcat, errorIfNULL=TRUE)

\S4method{features}{ANY}(siamcat)

\S4method{features}{otu_table}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link[phyloseq]{otu_table-class} .}
}
\value{
The \link[phyloseq]{otu_table-class} object or NULL.
}
\description{
Retrieve a \link[phyloseq]{otu_table-class} object from object.
}
\examples{
    data(siamcat_example)
    features(siamcat_example)
}
