% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{filt_params}
\alias{filt_params}
\alias{filt_params,ANY-method}
\title{Retrieve the list of filtering parameters from object.}
\usage{
filt_params(siamcat, verbose=1)

\S4method{filt_params}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a filt_feat object}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The list of filtering parameters or NULL.
}
\description{
Retrieve the list of filtering parameters from object.
}
\examples{
    data(siamcat_example)
    filt_params(siamcat_example)
}
