% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{filter.feat}
\alias{filter.feat}
\title{This function filters features in an unsupervised manner.}
\usage{
filter.feat(feat, filter.method, cutoff, recomp.prop, rm.unmapped)
}
\arguments{
\item{filter.method}{TODO}

\item{recomp.prop}{boolean, should absolute abundances be converted into relative abundances?}

\item{rm.unmapped}{TODO}

\item{fn.in.meta}{meta-data input file}

\item{fn.out.meta}{meta-data output file}

\item{abundance.cutoff}{TODO}
}
\description{
This function filters features in an unsupervised manner.
}
\examples{
filter.feat()
}
\keyword{SIAMCAT}
\keyword{association_check}
