% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.features.r
\name{filter.features}
\alias{filter.features}
\title{Perform unsupervised feature filtering.}
\usage{
filter.features(siamcat, filter.method = "abundance", cutoff = 0.001,
  recomp.prop = FALSE, rm.unmapped = TRUE, verbose = 1)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat}}

\item{filter.method}{method used for filtering the features, can be one of
these: \code{c("abundance", "cum.abundance", "prevalence")}}

\item{cutoff}{float, abundace or prevalence cutoff}

\item{recomp.prop}{boolean, should absolute abundances be
converted into relative abundances?}

\item{rm.unmapped}{boolean, should unmapped reads be discarded?}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for standard information, defaults to \code{1}}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for only information about progress and success, \code{2} for normal 
level of information and \code{3} for full debug information, defaults to \code{1}}
}
\value{
siamcat an object of class \link{siamcat}
}
\description{
This function may convert absolute abundances into relative
       abundances and then performs unsupervised feature filtering. Features
       can be filtered based on abundance or prevalence. Additionally,
       unmapped reads may be removed.
}
\details{
This function
}
\keyword{SIAMCAT}
\keyword{filter.features}
