% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{filter.label}
\alias{filter.label}
\title{Filter samples from \code{siamcat@label}}
\usage{
filter.label(siamcat, ids, verbose = 1)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat}}

\item{ids}{names of samples to be left in the \code{siamcat@label}}
}
\value{
siamcat an object of class \link{siamcat}
}
\description{
This functions filters \code{siamcat@label}.
}
\keyword{filter.label}
