% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_helper_functions.R
\name{filter.label}
\alias{filter.label}
\title{Filter samples from \code{siamcat@label}}
\usage{
filter.label(siamcat, ids, verbose = 1)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}

\item{ids}{names of samples to be left in the \code{siamcat@label}}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for more information about progress and success, defaults to \code{1}}
}
\value{
siamcat an object of class \link{siamcat-class}
}
\description{
This functions filters \code{siamcat@label}.
}
\examples{

 data(siamcat_example)
 # simple working example
 siamcat_filtered <- filter.label(siamcat_example, ids=c(1:10))

}
\keyword{filter.label}
