% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{get.features}
\alias{get.features}
\alias{get.features}
\alias{get.phyloseq}
\title{get.features}
\usage{
get.features(siamcat)

get.phyloseq(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}

\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
Features as \link[phyloseq]{otu_table} object

Object of class \link[phyloseq]{phyloseq-class}
}
\description{
Function to access features in siamcat@phylose@otu_table

Function to access phyloseq object in siamcat@phylose
}
\details{
Access features in siamcat@phylose@otu_table

Access phyloseq object in siamcat@phylose
}
\examples{
 data(siamcat_example)
 feat <- get.features(siamcat)
 data(siamcat_example)
 phyloseq <- get.phyloseq(siamcat)
}
