% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.filt_feat.matrix}
\alias{get.filt_feat.matrix}
\title{get.filt_feat.matrix}
\usage{
get.filt_feat.matrix(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
Filtered features as a matrix
}
\description{
Function to access features in siamcat@filt_feat@filt.feat
}
\details{
Access features in siamcat@filt_feat@filt.feat as matrix
}
\examples{
    data(siamcat_example)
    feat <- get.filt_feat.matrix(siamcat_example)
}
