% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.label}
\alias{get.label}
\title{get.label}
\usage{
get.label(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}}
}
\value{
an object of class \link{label-class}
}
\description{
Function to access labels in siamcat@label
}
\details{
Access label object in siamcat@label
}
\examples{
 data(siamcat_example)
 label <- get.label(siamcat_example)
}
