% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\name{get.orig_feat.matrix}
\alias{get.orig_feat.matrix}
\title{get.orig_feat.matrix}
\usage{
get.orig_feat.matrix(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat-class}t}
}
\value{
Original features as a matrix
}
\description{
Function to access original features in siamcat@orig_feat
}
\details{
Access original features in siamcat@orig_feat as matrix
}
\examples{
 data(siamcat_example)
 orig_feat <- get.orig_feat.matrix(siamcat_example)
}
