% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{interpretor.model.plot}
\alias{interpretor.model.plot}
\title{model interpretation plot}
\usage{
interpretor.model.plot(fn.plot = "model_interpretation.pdf", feat, label,
  meta, model, pred, positive.lab, negative.lab, color.scheme, consens.thres)
}
