% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{label-class}
\alias{label-class}
\title{The S4 class for storing label info.}
\description{
The S4 class for storing label info.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{inherited from \code{\link{list}} class, contains
a list with:
\itemize{
    \item \code{label} numeric vector, specifying to which category samples
    belong, usualy made of 1s and -1s
    \item \code{header} contains information from the header of the label
    file
    \item \code{info} list with additional informations about the dataset
    \item \code{positive.lab} specifies which of two numbers in label is a
    positive label
    \item \code{negative.lab} specifies which of two numbers in label is a
    negative label
    \item \code{n.idx} numeric vector - on which positions in the label there
    are samples with negative label
    \item \code{p.idx} numeric vector - on which positions in the label there
    are samples with positive label
    \item \code{n.lab} character string with a name for the negative label
    (e.g. 'healthy')
    \item \code{p.lab} character string with a name for the positive label
    (e.g. 'cancer')}}
}}

