% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{label}
\alias{label}
\alias{label,ANY-method}
\alias{label,label-method}
\alias{label,list-method}
\title{Retrieve a \link{label-class} object from object.}
\usage{
label(siamcat, verbose=1)

\S4method{label}{ANY}(siamcat, verbose = 1)

\S4method{label}{label}(siamcat)

\S4method{label}{list}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link{label-class} or a list.}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The \link{label-class} object or NULL.
}
\description{
Retrieve a \link{label-class} object from object.
}
\examples{
    data(siamcat_example)
    label(siamcat_example)
}
