% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{label}
\alias{label}
\alias{label,ANY-method}
\alias{label,label-method}
\alias{label}
\alias{label,list-method}
\alias{label}
\alias{orig_feat,otu_table-method}
\alias{label}
\title{Retrieve a \link{label-class} object from object.}
\usage{
label(siamcat)

\S4method{label}{ANY}(siamcat)

\S4method{label}{label}(siamcat)

\S4method{label}{list}(siamcat)

\S4method{orig_feat}{otu_table}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link{label-class} or a list.}
}
\value{
The \link{label-class} object or NULL.
}
\description{
Retrieve a \link{label-class} object from object.
}
\examples{
    data(siamcat_example)
    label(siamcat_example)
}
