% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrn_liblinear.R
\encoding{UTF-8}
\name{LearnerClassifLiblineaR}
\alias{LearnerClassifLiblineaR}
\title{LiblineaR Classification Learner}
\description{
LiblineaR Classification Learner

LiblineaR Classification Learner
}
\details{
Type of SVC depends on \code{type} argument:
\itemize{
\item \code{0} L2-regularized logistic regression (primal)
\item \code{1} L2-regularized L2-loss support vector classification (dual)
\item \code{3} L2-regularized L1-loss support vector classification (dual)
\item \code{2} L2-regularized L2-loss support vector classification (primal)
\item \code{4} Support vector classification by Crammer and Singer
\item \code{5} L1-regularized L2-loss support vector classification
\item \code{6} L1-regularized logistic regression
\item \code{7} L2-regularized logistic regression (dual)
}
If number of records > number of features, \code{type = 2} is faster 
than  \code{type = 1}
(Hsu et al. 2003).

Note that probabilistic predictions are only available for 
types \code{0}, \code{6}, and \code{7}.
The default \code{epsilon} value depends on the \code{type} parameter, 
see [LiblineaR::LiblineaR].
}
\keyword{internal}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifLiblineaR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerClassifLiblineaR-new}{\code{LearnerClassifLiblineaR$new()}}
\item \href{#method-LearnerClassifLiblineaR-clone}{\code{LearnerClassifLiblineaR$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifLiblineaR-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifLiblineaR-new}{}}}
\subsection{Method \code{new()}}{
#' Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifLiblineaR$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifLiblineaR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifLiblineaR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifLiblineaR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
