% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{make.predictions}
\alias{make.predictions}
\title{Prediction on the test set}
\usage{
make.predictions(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for only information about progress and success, \code{2} for normal 
level of information and \code{3} for full debug information, defaults to \code{1}}
}
\value{
object of class \link{siamcat-class}
}
\description{
This function takes the test set instances and the model trained by \link{plm.trainer} in order to predict the classes.
}
\keyword{SIAMCAT}
\keyword{plm.predictor}
