% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,ANY-method}
\alias{meta,sample_data-method}
\title{Retrieve a \link[phyloseq]{sample_data-class} object from object.}
\usage{
meta(siamcat)

\S4method{meta}{ANY}(siamcat)

\S4method{meta}{sample_data}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link[phyloseq]{sample_data-class}.}
}
\value{
The \link[phyloseq]{sample_data-class} object or NULL.
}
\description{
Retrieve a \link[phyloseq]{sample_data-class} object from object.
}
\examples{
    data(siamcat_example)
    meta(siamcat_example)
}
