% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{meta.crc.zeller}
\alias{meta.crc.zeller}
\title{Example metadata matrix}
\source{
\url{http://msb.embopress.org/content/10/11/766}
}
\description{
Metadata (as data.frame) of the CRC dataset from Zeller et
al. MSB 2014 (see \url{http://msb.embopress.org/content/10/11/766}),
containing 6 metadata variables variables (e.g. Age or BMI) for 141 samples.
}
\keyword{data}
