% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation_plot.r
\name{model.evaluation.plot}
\alias{model.evaluation.plot}
\title{Model Evaluation Plot}
\usage{
model.evaluation.plot(..., fn.plot, colours = NULL, verbose = 1)
}
\arguments{
\item{...}{one or more object of class \link{siamcat-class}, can be named}

\item{fn.plot}{string, filename for the pdf-plot}

\item{colours}{colour specification for the different \link{siamcat-class}-
objects, defaults to \code{NULL} which will cause the colours to be
picked from the \code{'Set1'} palette}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for only information about progress and success, \code{2} for normal
level of information and \code{3} for full debug information,
defaults to \code{1}}
}
\value{
Does not return anything, but produces the model evaluation plot.
}
\description{
Produces two plots for model evaluation. The first plot shows
    the Receiver Operating Characteristic (ROC)-curves, the other the
    Precision-recall (PR)-curves for the different cross-validation
    repetitions.
}
\examples{

    data(siamcat_example)
    # simple working example
    model.evaluation.plot(siamcat_example, fn.plot='./eval,pdf')

}
\keyword{SIAMCAT}
\keyword{model.evaluation.plot}
