% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{model_list}
\alias{model_list}
\alias{model_list,ANY-method}
\alias{model_list,model_list-method}
\title{Retrieve model_list from object.}
\usage{
model_list(siamcat, errorIfNULL=TRUE)

\S4method{model_list}{ANY}(siamcat, errorIfNULL = TRUE)

\S4method{model_list}{model_list}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a model_list or instance of \link{model_list-class}.}

\item{errorIfNULL}{Should the accessor stop with 
an error if the slot is empty (\code{NULL})? Default \code{TRUE}.}
}
\value{
The \link{model_list-class} object or NULL.
 By default this method stops with an error in the latter case.
}
\description{
Retrieve model_list from object.
}
\examples{
 data(siamcat_example)
 model_list(siamcat_example)
}
