% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{model_list}
\alias{model_list}
\alias{model_list,ANY-method}
\alias{model_list,model_list-method}
\alias{models,model_list-method}
\alias{model_type,model_list-method}
\alias{weight_matrix,model_list-method}
\title{Retrieve \link{model_list-class} from object.}
\usage{
model_list(siamcat)

\S4method{model_list}{ANY}(siamcat, verbose = 1)

\S4method{model_list}{model_list}(siamcat)

\S4method{models}{model_list}(siamcat)

\S4method{model_type}{model_list}(siamcat, verbose = 1)

\S4method{weight_matrix}{model_list}(siamcat, verbose = 1)

\S4method{model_type}{model_list}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a model_list or instance of \link{model_list-class}.}
}
\value{
The \link{model_list-class} object or NULL.
}
\description{
Retrieve \link{model_list-class} from object.
}
\examples{
    data(siamcat_example)
    model_list(siamcat_example)
}
