% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{models}
\alias{models}
\alias{models,siamcat-method}
\title{Retrieve list of trained models from a SIAMCAT object}
\usage{
models(siamcat, verbose=1)

\S4method{models}{siamcat}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains trained models}

\item{verbose}{integer, if the slot is empty, should a message be printed?
values can be either \code{0} (no output) or \code{1} (print message)}
}
\value{
The list of models or \code{NULL}
}
\description{
Function to retrieve the list of trained models
}
\details{
The function extracts the list of trained models.
}
\examples{
data(siamcat_example)
temp <- models(siamcat_example)
temp[[1]]
}
