% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{models}
\alias{models}
\alias{models,ANY-method}
\title{Retrieve list of models from object.}
\usage{
models(siamcat, errorIfNULL=TRUE)

\S4method{models}{ANY}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a model_list or instance of \link{model_list-class}.}
}
\value{
The list of models or NULL.
}
\description{
Retrieve list of models from object.
}
\examples{
    data(siamcat_example)
    models(siamcat_example)
}
