% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{norm_param}
\alias{norm_param}
\alias{norm_param,ANY-method}
\alias{norm_param,list-method}
\alias{norm_param}
\title{Retrieve norm_param from object.}
\usage{
norm_param(siamcat)

\S4method{norm_param}{ANY}(siamcat)

\S4method{norm_param}{list}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a norm_param}
}
\value{
The norm_param list or NULL.
}
\description{
Retrieve norm_param from object.
}
\examples{
    data(siamcat_example)
    norm_param(siamcat_example)
}
