% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_feat.r
\name{normalize.feat}
\alias{normalize.feat}
\title{Perform feature normalization according to specified parameters.}
\usage{
normalize.feat(feat, norm.method, log.n0, sd.min.q, n.p, n.sample, n.feature,
  n.global)
}
\arguments{
\item{feat}{feature object}

\item{norm.method}{normalization method, can be one of these: \code{c("rank.unit", "rank.std", "log.std", "log.unit")}}

\item{log.n0}{pseudocount to be added before log-transformation}

\item{sd.min.q}{quantile of the distribution of standard deviation of all feature that will be added to the denominator during standardization of each feature in order to avoid underestimation (only for methods=\code{c("rank.std", "log.std")})}

\item{n.p}{vector norm to use, can be either \code{1} or \code{2}, (only for method=\code{"log.unit"})}

\item{n.sample}{boolean, normalize by feature? (only for method=\code{"log.unit"})}

\item{n.feature}{boolean, normalize by sample? (only for method=\code{"log.unit"})}

\item{n.global}{boolean, Normalize by global rescaling? (only for method=\code{"log.unit"})}
}
\value{
list containing the matrix of normalized features and a list of normalization parameters: \itemize{
 \item \code{$par} = parameters utilized in the normalization;
 \item \code{$feat} = normalized features
}
}
\description{
This function performs feature normalization according to user-specified parameters.
}
\details{
There are four different normalization methods implemented:
\itemize{
 \item \code{"rank.unit"} converts features to ranks and normalizes each column by the square root of the sum of ranks
 \item \code{"rank.std"} converts features to ranks and applies z-score standardization
 \item \code{"log.unit"} log-transforms features (after addition of pseudocounts), converts features to ranks and normalizes by features or samples
 \item \code{"log.std"} log-transforms features (after addition of pseudocounts) and applies z-score standardization
}
The other parameters are depending on the normalization method of choice, e.g. \code{log.n0} as pseudocount is only needed for the methods making use of the log-transformation.
}
\keyword{SIAMCAT}
\keyword{normalize.feat}
