% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{orig_feat}
\alias{orig_feat}
\alias{orig_feat,siamcat-method}
\alias{orig_feat,otu_table-method}
\title{Retrieve a \link[phyloseq]{otu_table-class} object from otu_table slot in
the phyloseq slot in a siamcat object}
\usage{
orig_feat(siamcat)

\S4method{orig_feat}{siamcat}(siamcat)

\S4method{orig_feat}{otu_table}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a label or instance of \link[phyloseq]{otu_table-class}.}
}
\value{
The \link[phyloseq]{otu_table-class} object or NULL.
}
\description{
Retrieve a \link[phyloseq]{otu_table-class} object from otu_table slot in
the phyloseq slot in a siamcat object
}
\examples{
data(siamcat_example)
temp <- orig_feat(siamcat_example)
}
\keyword{internal}
