% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_trainer.r
\name{plm.trainer}
\alias{plm.trainer}
\title{Model training}
\usage{
plm.trainer(feat, label, cl = "lasso", data.split = NULL, stratify = TRUE,
  modsel.crit = "auc", min.nonzero.coeff = 1)
}
\arguments{
\item{feat}{features object}

\item{label}{label object}

\item{cl}{class of learner, directly passed to \link[mlr]{makeLearner}}

\item{data.split}{filename containing the training samples or list of training instances produced by \link{data.splitter()}, defaults to \code{NULL} leading to training on the complete dataset}

\item{stratify}{boolean, should the folds in the internal cross-validation be stratified?}

\item{min.nonzero.coeff}{integer number of minimum nonzero coefficients that should be present in the model}
}
\value{
an object of class \link[mlr]{makeWrappedModel}
}
\description{
This function trains the a machine learning model on the training data, using a \code{num.folds}-fold internal cross-validation scheme to find the optimal hyper-parameters of the model.
}
\keyword{SIAMCAT}
\keyword{plm.trainer}
