% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{pred_matrix}
\alias{pred_matrix}
\alias{pred_matrix,ANY-method}
\alias{pred_matrix,matrix-method}
\title{Retrieve pred_matrix from object.}
\usage{
pred_matrix(siamcat, verbose=1)

\S4method{pred_matrix}{ANY}(siamcat, verbose = 1)

\S4method{pred_matrix}{matrix}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a pred_matrix}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
The pred_matrix matrix or NULL.
}
\description{
Retrieve pred_matrix from object.
}
\examples{
    data(siamcat_example)
    pred_matrix(siamcat_example)
}
