% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{pred_matrix}
\alias{pred_matrix}
\alias{pred_matrix,ANY-method}
\alias{pred_matrix,matrix-method}
\alias{pred_matrix}
\title{Retrieve pred_matrix from object.}
\usage{
pred_matrix(siamcat)

\S4method{pred_matrix}{ANY}(siamcat)

\S4method{pred_matrix}{matrix}(siamcat)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a pred_matrix}
}
\value{
The pred_matrix matrix or NULL.
}
\description{
Retrieve pred_matrix from object.
}
\examples{
    data(siamcat_example)
    pred_matrix(siamcat_example)
}
