% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{read.features}
\alias{read.features}
\title{Read feature file}
\usage{
read.features(fn.in.feat)
}
\arguments{
\item{fn.in.feat}{name of the tsv file containing features}
}
\value{
matrix containing features from the file
}
\description{
This file reads in the tsv file with features and converts it into a matrix
}
