% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.r
\name{read.features}
\alias{read.features}
\title{Read feature file}
\usage{
read.features(fn.in.feat, transpose = FALSE, verbose = 0)
}
\arguments{
\item{fn.in.feat}{name of the tsv file containing features}

\item{transpose}{should the features table be transposed?}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for information about progress and time, defaults to \code{0}}
}
\value{
\code{otu_table} containing features from the file
}
\description{
This file reads in the tsv file with features and
converts it into a matrix.

The file should be oragnized as follows:
features (in rows) x samples (in columns).

First row should contain sample labels (consistent with label data), while
the first column should contain feature labels (e.g. taxonomic identifiers).
The remaining entries are expected to be real values \code{>= 0} that
quantify the abundance of each feature in each sample.
}
\examples{
    # run with example data
    fn.feat <- system.file('extdata', 'feat_crc_zeller_msb_mocat_specI.tsv',
    package = 'SIAMCAT')

    features <- read.features(fn.feat)
}
