% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{read.labels}
\alias{read.labels}
\title{Read labels file}
\usage{
read.labels(fn.in.label, feat = NULL)
}
\arguments{
\item{fn.in.label}{name of the tsv file containing labels}
}
\value{
list with two values: - $header contains header of the file and - $label stores matrix containing features from the file
}
\description{
This file reads in the tsv file with labels and converts it into a matrix
}
