% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.r
\name{read.labels}
\alias{read.labels}
\title{Read labels file}
\usage{
read.labels(fn.in.label)
}
\arguments{
\item{fn.in.label}{name of the tsv file containing labels}
}
\value{
label object containing several entries:\itemize{
\item \code{$label} named vector containing the numerical labels from the
file;
\item \code{$header} first row of the label file;
\item \code{$info} information about the type of label (e.g. \code{BINARY});
\item \code{$positive.lab} numerical label for controls, e.g. \code{-1};
\item \code{$negative.lab} numerical label for cases, e.g. \code{1};
\item \code{$n.idx} logical vector of labels (\code{TRUE} for controls,
\code{FALSE} otherwise);
\item \code{$n.lab} label for controls, e.g. \code{healthy};
\item \code{$p.idx} logical vector of labels (\code{TRUE} for cases,
\code{FALSE} otherwise);
\item \code{$p.lab} label for cases, e.g. \code{cancer}
}
}
\description{
This file reads in the tsv file with labels and converts it
into a label object.

First row is expected to be \code{#BINARY:1=[label for cases];
-1=[label for controls]}.
Second row should contain the sample identifiers as tab-separated list
(consistent with feature and metadata).

Third row is expected to contain the actual class labels (tab-separated):
\code{1} for each case and \code{-1} for each control.

Note: Labels can take other numeric values (but not characters or strings);
importantly, the label for cases has to be greater than the one for controls
}
\examples{
    # run with example data
fn.label <- system.file('extdata', 'label_crc_zeller_msb_mocat_specI.tsv',
    package = 'SIAMCAT')

labels <- read.labels(fn.label)
}
