% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefseformat.R
\name{read.lefse}
\alias{read.lefse}
\title{read an input file in a LEfSe input format}
\usage{
read.lefse(filename = "data.txt", n.meta = 1)
}
\arguments{
\item{filename}{name of the input file in a LEfSe input format}
}
\description{
This reads an input file in a LEfSe input format
}
\keyword{read.lefse}
