% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_utils.r
\name{read.meta}
\alias{read.meta}
\title{Read metadata file}
\usage{
read.meta(fn.in.meta)
}
\arguments{
\item{fn.in.meta}{name of the tsv file containing metadata}
}
\value{
matrix containing metadata from the file
}
\description{
This file reads in the tsv file with numerical metadata and converts it into a matrix.

The file should be organized as follows:
samples (in rows) x metadata (in columns). Metadata needs to be converted to numerical values by the user.

Metadata may be optional for the SIAMCAT workflow, but are necessary for heatmap displays, see \link{interpretor.model.plot}
}
