% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_samples.r
\name{select.samples}
\alias{select.samples}
\title{Select samples based on siamcat@phyloseq@sam_datadata}
\usage{
select.samples(siamcat, filter, allowed.set = NULL, allowed.range = NULL)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat}}

\item{filter}{name of the column from siamcat@phyloseq@sam_datadata on which the selection should be done}

\item{allowed.set}{a vector of allowed values}

\item{allowed.range}{a range of allowed values}
}
\value{
siamcat an object of class \link{siamcat}
}
\description{
This functions filters features, labels, and siamcat@phyloseq@sam_datadata based on a specific column in the siamcat@phyloseq@sam_datadata.
Provided with a column in the siamcat@phyloseq@sam_datadata and a range or a set of allowed values, the function will return the labels, siamcat@phyloseq@sam_datadata, and features for the samples matching the allowed range or set.
}
\keyword{SIAMCAT}
\keyword{select.samples}
