% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{select.samples}
\alias{select.samples}
\title{Validate data}
\usage{
select.samples(meta, feat, label, filter, allowed.set = NULL,
  allowed.range = NULL)
}
\arguments{
\item{meta}{metadata object}

\item{feat}{features object}

\item{label}{labels object}

\item{filter}{name of the column from metadata on which the selection should be done}

\item{allowed.set}{a vector of allowed values}

\item{allowed.range}{a range of allowed values}
}
\value{
a list containing values after filtering: $feat - features; $label - labels; $meta - metadata
}
\description{
Filter samples based on metadata
}
