% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{select.samples}
\alias{select.samples}
\title{Select samples}
\usage{
select.samples(meta, feat, label, filter, allowed.set = NULL,
  allowed.range = NULL)
}
\arguments{
\item{meta}{metadata object}

\item{feat}{features object}

\item{label}{labels object}

\item{filter}{name of the column from metadata on which the selection should be done}

\item{allowed.set}{a vector of allowed values}

\item{allowed.range}{a range of allowed values}
}
\value{
list containing values after selection: \itemize{
 \item \code{$feat} = features;
 \item \code{$label} = labels;
 \item \code{$meta} = metadata
}
}
\description{
This functions filters samples based on metadata.
Provided with a column in the metadata and a range or a set of allowed values, the function will return the labels, metadata, and features for the samples matching the allowed range or set.
}
\keyword{SIAMCAT}
\keyword{select.samples}
