% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_helper_functions.R
\name{show,siamcat-method}
\alias{show,siamcat-method}
\title{Show method for siamcat class object}
\usage{
\S4method{show}{siamcat}(object)
}
\value{
none
}
\description{
Show method for siamcat class object
}
\keyword{internal}
