% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class.R
\docType{class}
\name{siamcat-class}
\alias{siamcat-class}
\title{The S4 class for storing taxa-abundance information and models.}
\description{
The S4 class for storing taxa-abundance information and models.
}
\section{Slots}{

\describe{
\item{\code{phyloseq}}{object of class \link[phyloseq]{phyloseq-class}}

\item{\code{label}}{an object of class \link{label-class}}

\item{\code{orig_feat}}{an object of class \link[phyloseq]{otu_table-class}}

\item{\code{data_split}}{an object of class \link{data_split-class}}

\item{\code{norm_param}}{a list of normalzation parameters, see \link{normalize.features} for more details}

\item{\code{model_list}}{an object of class \link{model_list-class}}

\item{\code{eval_data}}{list containing \itemize{
 \item \code{$roc.average} average ROC-curve across repeats or a single ROC-curve on complete dataset;
 \item \code{$auc.average} AUC value for the average ROC-curve;
 \item \code{$ev.list} list of \code{length(num.folds)}, containing for different decision thresholds the number of false positives, false negatives, true negatives, and true positives;
 \item \code{$pr.list} list of \code{length(num.folds)}, containing the positive predictive value (precision) and true positive rate (recall) values used to plot the PR curves;
}. If \code{prediction} had more than one column, i.e. if the models has been trained with several repeats, the function will additonally return \itemize{
 \item \code{$roc.all} list of roc objects (see \link[pROC]{roc}) for every repeat;
 \item \code{$aucspr} vector of AUC values for the PR curves for every repeat;
 \item \code{$auc.all} vector of AUC values for the ROC curves for every repeat
}}

\item{\code{pred_matrix}}{a matrix with predictions made by \link{make.predictions} function}
}}

