% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_all_class_definitions.R
\docType{class}
\name{siamcat-class}
\alias{siamcat-class}
\title{The S4 class for storing taxa-abundance information and models.}
\description{
The S4 class for storing taxa-abundance information and models.
}
\section{Slots}{

\describe{
\item{\code{phyloseq}}{object of class \link[phyloseq]{phyloseq-class}}

\item{\code{label}}{an object of class \link{label-class}}

\item{\code{filt_feat}}{an object of class \link{filt_feat-class}}

\item{\code{associations}}{an object of class \link{associations-class}}

\item{\code{norm_feat}}{an object of class \link{norm_feat-class}}

\item{\code{data_split}}{an object of class \link{data_split-class}}

\item{\code{model_list}}{an object of class \link{model_list-class}}

\item{\code{eval_data}}{an object of class \link{eval_data-class}}

\item{\code{pred_matrix}}{an object of class \link{pred_matrix-class}}
}}

