% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class.R
\name{siamcat}
\alias{siamcat}
\title{Build siamcat-class objects from their components.}
\usage{
siamcat(...)
}
\arguments{
\item{...}{list of arguments needed in order to construct a SIAMCAT object}
}
\description{
Build siamcat-class objects from their components.
}
