% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_methods.R
\name{siamcat}
\alias{siamcat}
\title{siamcat}
\usage{
siamcat(...)
}
\arguments{
\item{...}{list of arguments needed in order to construct a SIAMCAT object}
}
\value{
A new \link{siamcat-class} object
}
\description{
Function to construct an object of class \link{siamcat-class}
}
\details{
Build siamcat-class objects from their components.
}
