% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_constructor.R
\name{siamcat}
\alias{siamcat}
\title{siamcat}
\usage{
siamcat(...)
}
\arguments{
\item{...}{list of arguments needed in order to construct a SIAMCAT object}
}
\value{
A new \link{siamcat-class} object
}
\description{
Function to construct an object of class \link{siamcat-class}
}
\details{
Build siamcat-class objects from their components.
}
\examples{
# example with package data
fn.in.feat    <- system.file('extdata',
    'feat_crc_zeller_msb_mocat_specI.tsv',
    package = 'SIAMCAT')
fn.in.label <- system.file('extdata',
    'label_crc_zeller_msb_mocat_specI.tsv',
    package = 'SIAMCAT')
fn.in.meta    <- system.file('extdata',
'num_metadata_crc_zeller_msb_mocat_specI.tsv',
    package = 'SIAMCAT')

feat    <- read.features(fn.in.feat)
label <- read.labels(fn.in.label)
meta    <- read.meta(fn.in.meta)
siamcat <- siamcat(feat, label, meta)
}
