% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{siamcat_example}
\alias{siamcat_example}
\title{Documentation for the example siamcat object in the data folder}
\description{
Reduced version of the CRC dataset in inst/extdata, containing 100 features
    (15 associated features at 5\% FDR in the original dataset and 85 random
    other features) and 141 samples, saved after the complete SIAMCAT pipeline
    has been run.
    Therefore, contains entries in every siamcat-object slot,
    e.g, \code{eval_data} or \code{data_split}. Mainly used for running
    the examples in the function documentation
}
\keyword{data}
