% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\encoding{UTF-8}
\name{siamcat_example}
\alias{siamcat_example}
\title{SIAMCAT example}
\source{
\url{http://msb.embopress.org/content/10/11/766}
}
\description{
Reduced version of the CRC dataset from Zeller et al. MSB 2014
(see \url{http://msb.embopress.org/content/10/11/766}), containing 100
features (15 associated features at 5\% FDR in the original dataset and
85 random other features) and 141 samples, saved after the complete SIAMCAT
pipelinehas been run.

Thus, the example dataset contains entries in every slot of the SIAMCAT
object (see \link{siamcat-class}), e.g, \code{eval_data} or
\code{data_split}.

Mainly used for running the examples in the function documentation.
}
\keyword{data}
