% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_features.R
\name{summarize.features}
\alias{summarize.features}
\title{Summarize features}
\usage{
summarize.features(siamcat, level = "g__", keep_original_names = TRUE,
  verbose = 1)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for only information about progress and success, \code{2} for normal
level of information and \code{3} for full debug information,
defaults to \code{1}}

\item{siamcat}{at which level to summarize (g__ = genus)}
}
\value{
object of class \link{siamcat-class} with the
    slot \code{eval_data} filled
}
\description{
Summarize features
}
\examples{

    data(siamcat_example)
    # simple working example
    siamcat_evaluated <- summarize.features(siamcat_example)

}
\keyword{internal}
