% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{validate.data}
\alias{validate.data}
\title{This function checks for potential confounders in the metadata}
\usage{
validate.data(feat, label, meta = NULL)
}
\arguments{
\item{feat}{features object}

\item{label}{labels object}

\item{meta}{fmetadata object}
}
\value{
a list containing values after filtering: $feat - features; $label - labels; $meta - metadata
}
\description{
This function checks for potential confounders in the metadata
}
\keyword{SIAMCAT}
\keyword{validate.data}
