% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.r
\name{validate.data}
\alias{validate.data}
\title{Validate samples in siamcat@phyloseq@sam_datadata, labels, and features}
\usage{
validate.data(siamcat)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat}}
}
\value{
an object of class \link{siamcat}
}
\description{
This function checks if labels are available for all samples in features. Additionally validates siamcat@phyloseq@sam_datadata, if given
}
\keyword{SIAMCAT}
\keyword{validate.data}
