% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.r
\name{validate.data}
\alias{validate.data}
\title{Validate samples in siamcat@phyloseq@sam_datadata, labels, and features}
\usage{
validate.data(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{an object of class \link{siamcat}}

\item{verbose}{control output: \code{0} for no output at all, \code{1}
for only information about progress and success, \code{2} for normal 
level of information and \code{3} for full debug information, defaults to \code{1}}
}
\value{
an object of class \link{siamcat}
}
\description{
This function checks if labels are available for all samples in features. Additionally validates siamcat@phyloseq@sam_datadata, if given
}
\keyword{SIAMCAT}
\keyword{validate.data}
