% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.r
\name{validate.data}
\alias{validate.data}
\title{Validate samples in metadata, labels, and features}
\usage{
validate.data(feat, label, meta = NULL)
}
\arguments{
\item{feat}{features object}

\item{label}{labels object}

\item{meta}{metadata object}
}
\value{
a list containing values after filtering: $feat - features; $label - labels; $meta - metadata
}
\description{
Check if metadata is available for all samples in labels and vice versa. If yes, order metadata according to the order found in labels.
}
\keyword{SIAMCAT}
\keyword{validate.data}
