% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\encoding{UTF-8}
\name{volcano.plot}
\alias{volcano.plot}
\title{Visualize associations between features and classes as volcano plot}
\usage{
volcano.plot(siamcat, fn.plot=NULL, color.scheme="RdYlBu", 
annotate=5)
}
\arguments{
\item{siamcat}{object of class \link{siamcat-class}}

\item{fn.plot}{string, filename for the pdf-plot. If \code{fn.plot} is
\code{NULL}, the plot will be produced in the active graphics device.}

\item{color.scheme}{valid R color scheme or vector of valid R colors (must 
be of the same length as the number of classes), defaults to \code{'RdYlBu'}}

\item{annotate}{integer, number of features to annotate with the name}
}
\value{
Does not return anything, but produces a volcano plot based on
association measures
}
\description{
This function creates a volcano plot to vizualize the 
association between features and the label
}
\details{
bla bla bla
}
\examples{
# Example data
data(siamcat_example)

# Simple example
volcano.plot(siamcat_example, fn.plot='./volcano.pdf')
}
\keyword{SIAMCAT}
\keyword{volcano.plot}
