% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siamcat_class_accessors.R
\docType{methods}
\name{weight_matrix}
\alias{weight_matrix}
\alias{weight_matrix,ANY-method}
\title{Retrieve weight_matrix from object.}
\usage{
weight_matrix(siamcat, verbose=1)

\S4method{weight_matrix}{ANY}(siamcat, verbose = 1)
}
\arguments{
\item{siamcat}{(Required). An instance of \link{siamcat-class}
that contains a model_list or instance of \link{model_list-class}.}

\item{verbose}{If the slot is empty, should a message be printed? values
can be either 0 (no output) or 1 (print message)}
}
\value{
A matrix containing the feature weights or NULL
}
\description{
Retrieve weight_matrix from object.
}
\examples{
    data(siamcat_example)
    weight_matrix(siamcat_example)
}
